Initial Commit (Mathematical Approach).
In this problem, we have to find the minimum number of operations required to make the matrix beautiful. The matrix will be considered beautiful if the sum of all the columns and all the rows are equal. And the increment of 1, to the particular element in a matrix, will be considered as the single operation.
.
Time Complexity: O(n*n)
Auxiliary Space: O(n), where 'n' is the number of elements in a matrix.